S012 + correction

+++ src/main/java/fr/eni/ludotheque/dal/AdresseRepository.java
@@ -3,5 +3,5 @@ package fr.eni.ludotheque.dal;
 import fr.eni.ludotheque.bo.Adresse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdresseRepository extends JpaRepository<Adresse, Integer> {
+public interface AdresseRepository extends JpaRepository<Adresse, Long> {
 }
